pyproject.toml
[tool.poetry]
name = "simple_sitemap_parser"
version = "0.1.0"
description = ""
authors = ["Vjacheslav <goplit2010.konovalov@yandex.ru>"]
packages = [
    { include = "parser" },
]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
requests = "^2.27.1"
aiohttp = "^3.8.1"
python-dotenv = "^0.20.0"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
